Add testing for list, delete, help and extend delete command

+++ src/main/java/seedu/wildwatch/command/DeleteCommand.java
@@ -3,6 +3,8 @@ package seedu.wildwatch.command;
 import seedu.wildwatch.entry.EntryList;
 import seedu.wildwatch.operation.Ui;
 
+import java.util.Scanner;
+
 /**
  * Command class for deleting entry in EntryList
  */
@@ -20,8 +22,23 @@ public class DeleteCommand extends Command {
      * Deletes entry in the EntryList
      */
     public void execute() {
-        Ui.entryRemovedMessagePrinter();
-        EntryList.deleteEntry(numberInput);
-        Ui.entryCountPrinter();
+        if (numberInput <= 0 || numberInput > EntryList.getArraySize()) {
+            Ui.entryNotFoundMessagePrinter();
+        } else {
+            Scanner scanner = new Scanner(System.in);
+            Ui.printEntry(numberInput-1);
+            Ui.entryDeletionConfirmationPrinter();
+            String confirmation = scanner.nextLine().trim().toLowerCase();
+
+            if (confirmation.equals("yes")) {
+                Ui.entryRemovedMessagePrinter();
+                EntryList.deleteEntry(numberInput);
+                Ui.entryCountPrinter();
+            } else {
+                Ui.printHorizontalLines();
+                Ui.entryNotDeletedMessagePrinter();
+
+            }
+        }
     }
 }
